requirements.txt
selenium==4.30.0
undetected-chromedriver==3.5.5
# Core data processing
pandas==2.1.4
numpy==1.26.4
Unidecode==1.3.8
pyarrow>=15.0.2
openpyxl>=3.1.2
xlrd>=2.0.1
python-dateutil>=2.9.0

# Database connectivity
psycopg2-binary>=2.9.9
SQLAlchemy==1.4.46
pymysql>=1.1.0  # Optional: nếu dùng MySQL
pyodbc>=5.1.0   # Optional: nếu dùng SQL Server

# Big data processing
dask>=2024.4.1  # Optional: nếu xử lý dữ liệu lớn
# pyspark>=3.5.1  # Optional: nếu dùng Spark

# API and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.3
httpx>=0.27.0  # Optional: nếu cần HTTP async

# File and storage
boto3>=1.34.84
s3fs>=2024.3.1
google-cloud-storage>=2.16.0  # Optional: nếu dùng GCS
paramiko>=3.4.0  # Optional: nếu dùng SFTP

# Special formats
python-json-logger>=2.0.7
orjson>=3.10.0
fastparquet>=2024.2.0  # Optional: nếu cần Parquet nhanh hơn

# Airflow providers
apache-airflow-providers-common-sql>=1.13.0
apache-airflow-providers-http>=4.10.0
apache-airflow-providers-postgres>=5.10.3
apache-airflow-providers-amazon>=8.19.0  # Optional: nếu dùng AWS
apache-airflow-providers-google>=10.19.0  # Optional: nếu dùng Google Cloud

# Monitoring
prometheus-client>=0.20.0
sentry-sdk>=1.45.0  # Optional: nếu dùng Sentry